requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-google-translate==1.1
django-phonenumber-field==8.0.0
geographiclib==2.0
geopy==2.4.1
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
joblib==1.4.2
numpy==2.0.0
pandas==2.2.2
phonenumbers==8.13.40
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
